Feat: Can save student exam

+++ database/migrations/2023_11_01_001012_create_student_task_performances_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
 	        $table->uuid('id')->primary();
 	        $table->uuid('stpc_id')->index();
 	        $table->uuid('student_admission_id')->index();
-	        $table->float('score');
+	        $table->integer('score');
 	        
 	        $table->timestamps();
